Trait connecteer_capabilities::connection::Connection

pub trait Connection<Payload: Serialize + DeserializeOwned>: Sealed<Payload> {
    type Wrapped: Serialize + DeserializeOwned;
    type Ctx;
    type SendError;
    type ReceiveError;
    type NextError;
    type SendGen<'s, 'c, 'g>: Generator<(&'s mut Self, &'c mut Self::Ctx), Yield = Result<Self::Wrapped, Self::SendError>, Return = ()> + 'g
       where Self: 's,
             Self::Ctx: 'c;
    type ReceiveGen<'s, 'c, 'g>: Generator<(&'s mut Self, &'c mut Self::Ctx), Yield = Result<Self::Wrapped, Self::ReceiveError>, Return = ()> + 'g
       where Self: 's,
             Self::Ctx: 'c;

    // Required methods
    fn send<'s, 'c, 'g>(
        input: Payload,
        _: PublicUncallable
    ) -> Self::SendGen<'s, 'c, 'g>
       where Self: 's,
             Self::Ctx: 'c;
    fn receive<'s, 'c, 'g>(
        output: Result<Payload, Self::NextError>,
        _: PublicUncallable
    ) -> Self::ReceiveGen<'s, 'c, 'g>
       where Self::Ctx: 'c,
             Self: 's;
}

You can’t implement this trait, you need to let the blanket implementation do its job by implementing Middleware on your types This type isn’t used directly by the consumer, it is only used by this crate

Required Associated Types

type Wrapped: Serialize + DeserializeOwned

type Ctx

type SendError

type ReceiveError

type NextError

type SendGen<'s, 'c, 'g>: Generator<(&'s mut Self, &'c mut Self::Ctx), Yield = Result<Self::Wrapped, Self::SendError>, Return = ()> + 'g where Self: 's, Self::Ctx: 'c

type ReceiveGen<'s, 'c, 'g>: Generator<(&'s mut Self, &'c mut Self::Ctx), Yield = Result<Self::Wrapped, Self::ReceiveError>, Return = ()> + 'g where Self: 's, Self::Ctx: 'c

Required Methods

fn send<'s, 'c, 'g>( input: Payload, _: PublicUncallable ) -> Self::SendGen<'s, 'c, 'g>where Self: 's, Self::Ctx: 'c,

fn receive<'s, 'c, 'g>( output: Result<Payload, Self::NextError>, _: PublicUncallable ) -> Self::ReceiveGen<'s, 'c, 'g>where Self::Ctx: 'c, Self: 's,