Struct connecteer_capabilities::identity::Wrapper

#[repr(transparent)]
pub struct Wrapper<T>(pub T);

This type can be used when implementing an Middleware that doesn’t modify the message, but only act with side effects (for example logging message, or filtering)

Tuple Fields
0: T
Implementations

impl<T> Wrapper<T>

pub fn into_inner(self) -> T