Trait connecteer_capabilities::Middleware

pub trait Middleware<Payload: Serialize + DeserializeOwned> {
    type Message: Serialize + DeserializeOwned;
    type WrapError;
    type UnwrapError;
    type Ctx;
    type Next: Connection<Self::Message>;
    type WrapGen<'s, 'c, 'g>: Generator<(&'s mut Self, &'c mut Self::Ctx), Yield = Result<Self::Message, Self::WrapError>, Return = ()> + 'g
       where Self::Ctx: 'c,
             Self: 's;
    type UnwrapGen<'s, 'c, 'g>: Generator<(&'s mut Self, &'c mut Self::Ctx), Yield = Result<Payload, Self::UnwrapError>, Return = ()> + 'g
       where Self: 's,
             Self::Ctx: 'c;

    // Required methods
    fn wrap<'a, 'b, 'g, Uncallable: PublicUncallable>(
        msg: Payload
    ) -> Self::WrapGen<'a, 'b, 'g>;
    fn unwrap<'a, 'b, 'g, Uncallable: PublicUncallable>(
        msg: Self::Message
    ) -> Self::UnwrapGen<'a, 'b, 'g>;
    fn create_wrap_error<Uncallable: PublicUncallable>(
        &mut self,
        err: <Self::Next as Connection<Self::Message>>::SendError
    ) -> Self::WrapError;
    fn create_unwrap_error<Uncallable: PublicUncallable>(
        &mut self,
        err: <Self::Next as Connection<Self::Message>>::ReceiveError
    ) -> Self::UnwrapError;
    fn create_unwrap_error_passthrough<Uncallable: PublicUncallable>(
        &mut self,
        err: <Self::Next as Connection<Self::Message>>::NextError
    ) -> Self::UnwrapError;
    fn get_next<Uncallable: PublicUncallable>(&mut self) -> &mut Self::Next;
    fn get_next_ctx<Uncallable: PublicUncallable>(
        c: &mut Self::Ctx
    ) -> &mut <Self::Next as Connection<Self::Message>>::Ctx;
}

Required Associated Types

type Message: Serialize + DeserializeOwned

This is the message type that is outputted by the middleware when sending messages (and inputted when receiving messages)

type WrapError

The error type returned when wrapping an Message

type UnwrapError

The error type returned when unwrapping an Message and also provide a way to “passthrough” errors made by middleware after them

type Ctx

type Next: Connection<Self::Message>

type WrapGen<'s, 'c, 'g>: Generator<(&'s mut Self, &'c mut Self::Ctx), Yield = Result<Self::Message, Self::WrapError>, Return = ()> + 'g where Self::Ctx: 'c, Self: 's

type UnwrapGen<'s, 'c, 'g>: Generator<(&'s mut Self, &'c mut Self::Ctx), Yield = Result<Payload, Self::UnwrapError>, Return = ()> + 'g where Self: 's, Self::Ctx: 'c

Required Methods

fn wrap<'a, 'b, 'g, Uncallable: PublicUncallable>( msg: Payload ) -> Self::WrapGen<'a, 'b, 'g>

Transform an Message into a Unwrapped Payload

fn unwrap<'a, 'b, 'g, Uncallable: PublicUncallable>( msg: Self::Message ) -> Self::UnwrapGen<'a, 'b, 'g>

Transform an Payload into a Wrapped Message

fn create_wrap_error<Uncallable: PublicUncallable>( &mut self, err: <Self::Next as Connection<Self::Message>>::SendError ) -> Self::WrapError

This function allows the system to bubble-up an error when wrapping a Message

fn create_unwrap_error<Uncallable: PublicUncallable>( &mut self, err: <Self::Next as Connection<Self::Message>>::ReceiveError ) -> Self::UnwrapError

This function allows the system to create an error when unwrapping a Message

fn create_unwrap_error_passthrough<Uncallable: PublicUncallable>( &mut self, err: <Self::Next as Connection<Self::Message>>::NextError ) -> Self::UnwrapError

This function allows the system to bubble-up an error

fn get_next<Uncallable: PublicUncallable>(&mut self) -> &mut Self::Next

fn get_next_ctx<Uncallable: PublicUncallable>( c: &mut Self::Ctx ) -> &mut <Self::Next as Connection<Self::Message>>::Ctx